Adaptations necessary for newer versions of my esp firmware.
The newer versions no longer drop new tcp connections when the maximum
number of tcp connections has been reached, instead they accept the
connection and then close the connection immediately. Unfortunately
that is a restriction of the esp sdk. Adapted the espif source to
recognise a connection that is dropped immediately as a failed
connection likewise a connection timeout and a refused connection.